# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Tarelli API
  description: >
    API REST para la aplicación de gestión personal de tareas Tarelli.
    No existen roles: cada usuario autenticado solo puede acceder a sus
    propios recursos (ownership). Las respuestas exitosas tienen la forma
    `{ "data": ..., "meta": ... }`; los errores tienen la forma
    `{ "error": { "code", "message", "details" } }`.
  version: '1.0.0'
servers:
  - url: http://localhost:4000/api
    description: Desarrollo local (Docker Compose)

tags:
  - name: Auth
  - name: Tareas
  - name: Categorías
  - name: Etiquetas

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Access token JWT de corta duración, enviado como header Authorization: Bearer <token>."

  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: La solicitud contiene datos inválidos.
            details:
              type: object

    Usuario:
      type: object
      properties:
        id: { type: string, format: uuid }
        nombre: { type: string }
        email: { type: string, format: email }
        timezone: { type: string, example: America/Bogota }
        ultimoLoginEn: { type: string, format: date-time, nullable: true }
        creadoEn: { type: string, format: date-time }

    Categoria:
      type: object
      properties:
        id: { type: string, format: uuid }
        usuario_id: { type: string, format: uuid }
        nombre: { type: string }
        creado_en: { type: string, format: date-time }
        actualizado_en: { type: string, format: date-time }

    Etiqueta:
      type: object
      properties:
        id: { type: string, format: uuid }
        usuario_id: { type: string, format: uuid }
        nombre: { type: string }
        creado_en: { type: string, format: date-time }

    Tarea:
      type: object
      properties:
        id: { type: string, format: uuid }
        usuario_id: { type: string, format: uuid }
        categoria_id: { type: string, format: uuid, nullable: true }
        categoria_nombre: { type: string, nullable: true }
        titulo: { type: string }
        descripcion: { type: string }
        prioridad: { type: string, enum: [baja, media, alta] }
        completada: { type: boolean }
        fecha_vencimiento: { type: string, format: date, nullable: true }
        completado_en: { type: string, format: date-time, nullable: true }
        creado_en: { type: string, format: date-time }
        actualizado_en: { type: string, format: date-time }
        etiquetas:
          type: array
          items: { $ref: '#/components/schemas/Etiqueta' }

    TareaInput:
      type: object
      required: [titulo]
      properties:
        titulo: { type: string, maxLength: 200 }
        descripcion: { type: string, maxLength: 5000 }
        prioridad: { type: string, enum: [baja, media, alta], default: media }
        categoria_id: { type: string, format: uuid, nullable: true }
        fecha_vencimiento: { type: string, format: date, nullable: true }
        etiquetas:
          type: array
          items: { type: string, format: uuid }

    PaginatedMeta:
      type: object
      properties:
        page: { type: integer, example: 1 }
        limit: { type: integer, example: 20 }
        total: { type: integer, example: 42 }
        totalPages: { type: integer, example: 3 }

  responses:
    NoAutenticado:
      description: Falta o es inválido el access token.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NoEncontrado:
      description: El recurso no existe o no pertenece al usuario autenticado.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    ValidacionInvalida:
      description: Los datos enviados no pasan las validaciones.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    Conflicto:
      description: Conflicto con el estado actual (p. ej. nombre duplicado).
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }

paths:
  /auth/registro:
    post:
      tags: [Auth]
      summary: Registrar un nuevo usuario
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre, email, password]
              properties:
                nombre: { type: string, minLength: 2, maxLength: 100 }
                email: { type: string, format: email }
                password:
                  type: string
                  minLength: 8
                  description: Debe incluir al menos una letra y un número.
      responses:
        '201':
          description: Usuario creado y sesión iniciada.
          headers:
            Set-Cookie:
              description: Cookies `tarelli_refresh` (HttpOnly) y `tarelli_csrf`.
              schema: { type: string }
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      usuario: { $ref: '#/components/schemas/Usuario' }
                      accessToken: { type: string }
                  meta: { type: object }
        '422': { $ref: '#/components/responses/ValidacionInvalida' }
        '409': { $ref: '#/components/responses/Conflicto' }

  /auth/login:
    post:
      tags: [Auth]
      summary: Iniciar sesión
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Sesión iniciada.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      usuario: { $ref: '#/components/schemas/Usuario' }
                      accessToken: { type: string }
        '401':
          description: Credenciales inválidas (código `INVALID_CREDENTIALS`).
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '422': { $ref: '#/components/responses/ValidacionInvalida' }

  /auth/refresh:
    post:
      tags: [Auth]
      summary: Rota el refresh token (cookie) y emite un nuevo access token
      description: >
        Se autentica exclusivamente por la cookie `tarelli_refresh` (HttpOnly).
        Requiere el header `X-CSRF-Token` con el valor de la cookie
        `tarelli_csrf` (patrón de doble envío). Si el refresh token ya fue
        rotado previamente (reutilización), se revoca toda la sesión del
        usuario y responde 401.
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Nuevo access token y refresh token (rotado).
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      usuario: { $ref: '#/components/schemas/Usuario' }
                      accessToken: { type: string }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '403':
          description: Token CSRF inválido o ausente.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }

  /auth/logout:
    post:
      tags: [Auth]
      summary: Cierra la sesión (revoca el refresh token actual)
      parameters:
        - name: X-CSRF-Token
          in: header
          required: true
          schema: { type: string }
      responses:
        '204': { description: Sesión cerrada. }
        '403':
          description: Token CSRF inválido o ausente.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }

  /auth/perfil:
    get:
      tags: [Auth]
      summary: Perfil del usuario autenticado
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Perfil del usuario.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      usuario: { $ref: '#/components/schemas/Usuario' }
        '401': { $ref: '#/components/responses/NoAutenticado' }

  /tareas:
    get:
      tags: [Tareas]
      summary: Lista las tareas del usuario autenticado (filtros, búsqueda, orden, paginación)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: completada, in: query, schema: { type: boolean } }
        - { name: categoria, in: query, schema: { type: string, format: uuid } }
        - { name: prioridad, in: query, schema: { type: string, enum: [baja, media, alta] } }
        - { name: fecha_vencimiento, in: query, schema: { type: string, format: date } }
        - { name: fecha_vencimiento_desde, in: query, schema: { type: string, format: date } }
        - { name: fecha_vencimiento_hasta, in: query, schema: { type: string, format: date } }
        - { name: busqueda, in: query, schema: { type: string }, description: Full-Text Search (título/descripción). }
        - name: etiquetas
          in: query
          schema: { type: string }
          description: Lista de IDs de etiqueta separados por coma. Semántica AND — la tarea debe tener TODAS.
        - { name: ordenar, in: query, schema: { type: string, enum: [creado_en, fecha_vencimiento, prioridad, titulo] } }
        - { name: direccion, in: query, schema: { type: string, enum: [asc, desc] } }
        - { name: page, in: query, schema: { type: integer, minimum: 1, default: 1 } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        '200':
          description: Lista paginada de tareas.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Tarea' }
                  meta: { $ref: '#/components/schemas/PaginatedMeta' }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '422':
          description: ordenar/direccion fuera de la whitelist permitida, u otro parámetro inválido.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }

    post:
      tags: [Tareas]
      summary: Crea una tarea
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TareaInput' } } }
      responses:
        '201':
          description: Tarea creada.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: '#/components/schemas/Tarea' } }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '422':
          description: >
            Datos inválidos, o categoria_id/etiquetas que no existen o no
            pertenecen al usuario autenticado.
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }

  /tareas/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Tareas]
      summary: Actualización completa de una tarea (incluye la lista completa de etiquetas)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/TareaInput' } } }
      responses:
        '200':
          description: Tarea actualizada.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: '#/components/schemas/Tarea' } }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '422': { $ref: '#/components/responses/ValidacionInvalida' }
    delete:
      tags: [Tareas]
      summary: Elimina una tarea (hard delete; limpia relaciones y activity_logs en cascada)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Tarea eliminada. }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }

  /tareas/{id}/completar:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [Tareas]
      summary: Marca una tarea como completada o pendiente (idempotente)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [completada]
              properties: { completada: { type: boolean } }
      responses:
        '200':
          description: Tarea actualizada.
          content:
            application/json:
              schema:
                type: object
                properties: { data: { $ref: '#/components/schemas/Tarea' } }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '422': { $ref: '#/components/responses/ValidacionInvalida' }

  /categorias:
    get:
      tags: [Categorías]
      summary: Lista las categorías del usuario autenticado
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lista de categorías.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Categoria' } }
        '401': { $ref: '#/components/responses/NoAutenticado' }
    post:
      tags: [Categorías]
      summary: Crea una categoría
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre]
              properties: { nombre: { type: string, maxLength: 100 } }
      responses:
        '201':
          description: Categoría creada.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Categoria' } } }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '409': { $ref: '#/components/responses/Conflicto' }
        '422': { $ref: '#/components/responses/ValidacionInvalida' }

  /categorias/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    put:
      tags: [Categorías]
      summary: Renombra una categoría
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre]
              properties: { nombre: { type: string, maxLength: 100 } }
      responses:
        '200':
          description: Categoría actualizada.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Categoria' } } }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }
        '409': { $ref: '#/components/responses/Conflicto' }
    delete:
      tags: [Categorías]
      summary: Elimina una categoría (las tareas quedan con categoria_id = null)
      security: [{ bearerAuth: [] }]
      responses:
        '204': { description: Categoría eliminada. }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '404': { $ref: '#/components/responses/NoEncontrado' }

  /etiquetas:
    get:
      tags: [Etiquetas]
      summary: Lista las etiquetas del usuario autenticado
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Lista de etiquetas.
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Etiqueta' } }
        '401': { $ref: '#/components/responses/NoAutenticado' }
    post:
      tags: [Etiquetas]
      summary: Crea una etiqueta
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nombre]
              properties: { nombre: { type: string, maxLength: 50 } }
      responses:
        '201':
          description: Etiqueta creada.
          content:
            application/json:
              schema: { type: object, properties: { data: { $ref: '#/components/schemas/Etiqueta' } } }
        '401': { $ref: '#/components/responses/NoAutenticado' }
        '409': { $ref: '#/components/responses/Conflicto' }
        '422': { $ref: '#/components/responses/ValidacionInvalida' }